+++ crates/swc_css_codegen/Cargo.toml
@@ -21,6 +21,7 @@ swc_atoms              = { version = "0.4.18", path = "../swc_atoms" }
 swc_common             = { version = "0.29.3", path = "../swc_common" }
 swc_css_ast            = { version = "0.114.3", path = "../swc_css_ast" }
 swc_css_codegen_macros = { version = "0.2.0", path = "../swc_css_codegen_macros" }
+swc_css_utils          = { version = "0.111.1", path = "../swc_css_utils" }
 
 [dev-dependencies]
 swc_common = { version = "0.29.3", path = "../swc_common", features = [
